Attempt to fix things I don't understand - Target netstandard1.3 instead of 1.1 for things like the execution context. - Used System.Threading.Channels as a template
